add user login, authorization and testing

+++ part-04/01-bloglist/controllers/blogs.js
@@ -1,4 +1,5 @@
 const blogRouter = require('express').Router();
+const { userExtractor } = require('../utils/middlewares');
 const Blog = require('../models/Blog');
 const User = require('../models/User');
 
@@ -16,9 +17,10 @@ blogRouter.get('/:id', async (request, response, next) => {
   }
 });
 
-blogRouter.post('/', async (request, response, next) => {
+blogRouter.post('/', userExtractor, async (request, response, next) => {
   const body = request.body;
-  const user = await User.findById(body.userId);
+
+  const user = await User.findById(request.user.id);
 
   const blog = new Blog({
     ...body,
@@ -32,7 +34,7 @@ blogRouter.post('/', async (request, response, next) => {
   response.status(201).json(savedBlog);
 });
 
-blogRouter.put('/:id', async (request, response, next) => {
+blogRouter.put('/:id', userExtractor, async (request, response, next) => {
   const body = request.body;
   const updatedBlog = {
     likes: body.likes
@@ -42,8 +44,27 @@ blogRouter.put('/:id', async (request, response, next) => {
   response.status(200).json(updatedBlog);
 });
 
-blogRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndDelete(request.params.id);
+blogRouter.delete('/:id', userExtractor, async (request, response) => {
+  const user = request.user;
+  const blog = await Blog.findById(request.params.id);
+
+  if (blog === null) {
+    return response.status(404).json({ error: 'blog not found' });
+  }
+
+  if (!blog.user) {
+    return response.status(401).json({ error: 'unauthorized' });
+  }
+
+  if (blog.user.toString() !== user.id.toString()) {
+    return response.status(401).json({ error: 'unauthorized' });
+  }
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id);
+    return response.status(204).end();
+  }
+
   response.status(204).end();
 });
